🎨 remove some code and add more unit tests
