+++ query-understanding/nlp_tools/stopwords_filter.py
@@ -1,4 +1,3 @@
-# import spacy
 import nltk
 nltk.download('stopwords')
 from nltk.corpus import stopwords
@@ -7,16 +6,6 @@ STOPWORDS_ES = set(stopwords.words('spanish'))
 
 def remove_stopwords(words: list) -> list:
     
-    # nlp = spacy.load('es_core_news_sm')
-    
-    # filtered_sentence =[] 
-    # for word in words:
-    #     lexeme = nlp.vocab[word]
-    #     if lexeme.is_stop == False:
-    #         filtered_sentence.append(word) 
-    
-    # return filtered_sentence
-   
     no_stopwords = [ w  for w in words if w not in STOPWORDS_ES]
     return no_stopwords
