+++ query-understanding/nlp_tools/tokenizer.py
@@ -1,5 +1,5 @@
 import spacy
-import pickle
+
