--- query-understanding/normalization/lang_detec.py
@@ -1,4 +1,0 @@
-from textblob import TextBlob
-b = TextBlob("bonjour")
-b.detect_language()
-
